simplify task execution by removing middleware
