+++ cmd/uplift/bump.go
@@ -27,8 +27,6 @@ import (
 	"io"
 
 	"github.com/gembaadvantage/uplift/internal/context"
-	"github.com/gembaadvantage/uplift/internal/middleware/logging"
-	"github.com/gembaadvantage/uplift/internal/middleware/skip"
 	"github.com/gembaadvantage/uplift/internal/semver"
 	"github.com/gembaadvantage/uplift/internal/task"
 	"github.com/gembaadvantage/uplift/internal/task/bump"
@@ -110,7 +108,7 @@ func bumpFiles(opts bumpOptions, out io.Writer) error {
 		return err
 	}
 
-	tsks := []task.Runner{
+	tasks := []task.Runner{
 		gitcheck.Task{},
 		before.Task{},
 		gpgimport.Task{},
@@ -123,13 +121,7 @@ func bumpFiles(opts bumpOptions, out io.Writer) error {
 		after.Task{},
 	}
 
-	for _, tsk := range tsks {
-		if err := skip.Running(tsk.Skip, logging.Log(tsk.String(), tsk.Run))(ctx); err != nil {
-			return err
-		}
-	}
-
-	return nil
+	return task.Execute(ctx, tasks)
 }
 
 func setupBumpContext(opts bumpOptions, out io.Writer) (*context.Context, error) {
