+++ cmd/uplift/changelog.go
@@ -29,8 +29,6 @@ import (
 
 	"github.com/gembaadvantage/uplift/internal/context"
 	"github.com/gembaadvantage/uplift/internal/git"
-	"github.com/gembaadvantage/uplift/internal/middleware/logging"
-	"github.com/gembaadvantage/uplift/internal/middleware/skip"
 	"github.com/gembaadvantage/uplift/internal/task"
 	"github.com/gembaadvantage/uplift/internal/task/changelog"
 	"github.com/gembaadvantage/uplift/internal/task/gitcheck"
@@ -139,7 +137,7 @@ func writeChangelog(opts changelogOptions, out io.Writer) error {
 		return err
 	}
 
-	tsks := []task.Runner{
+	tasks := []task.Runner{
 		gitcheck.Task{},
 		before.Task{},
 		scm.Task{},
@@ -151,13 +149,7 @@ func writeChangelog(opts changelogOptions, out io.Writer) error {
 		after.Task{},
 	}
 
-	for _, tsk := range tsks {
-		if err := skip.Running(tsk.Skip, logging.Log(tsk.String(), tsk.Run))(ctx); err != nil {
-			return err
-		}
-	}
-
-	return nil
+	return task.Execute(ctx, tasks)
 }
 
 func writeChangelogDiff(opts changelogOptions, out io.Writer) error {
@@ -166,7 +158,7 @@ func writeChangelogDiff(opts changelogOptions, out io.Writer) error {
 		return err
 	}
 
-	tsks := []task.Runner{
+	tasks := []task.Runner{
 		gitcheck.Task{},
 		before.Task{},
 		scm.Task{},
@@ -174,13 +166,7 @@ func writeChangelogDiff(opts changelogOptions, out io.Writer) error {
 		after.Task{},
 	}
 
-	for _, tsk := range tsks {
-		if err := skip.Running(tsk.Skip, logging.Log(tsk.String(), tsk.Run))(ctx); err != nil {
-			return err
-		}
-	}
-
-	return nil
+	return task.Execute(ctx, tasks)
 }
 
 func setupChangelogContext(opts changelogOptions, out io.Writer) (*context.Context, error) {
