--- internal/middleware/logging/logging_test.go
@@ -1,49 +1,0 @@
-/*
-Copyright (c) 2022 Gemba Advantage
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-*/
-
-package logging
-
-import (
-	"testing"
-
-	"github.com/apex/log/handlers/cli"
-	"github.com/gembaadvantage/uplift/internal/context"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func TestLog(t *testing.T) {
-	err := Log("test", func(ctx *context.Context) error {
-		return nil
-	})(&context.Context{})
-
-	require.NoError(t, err)
-}
-
-func TestLog_PrettyPrints(t *testing.T) {
-	Log("test", func(ctx *context.Context) error {
-		require.Equal(t, 6, cli.Default.Padding)
-		return nil
-	})(&context.Context{})
-
-	assert.Equal(t, 3, cli.Default.Padding)
-}
