--- internal/middleware/skip/skip.go
@@ -1,45 +1,0 @@
-/*
-Copyright (c) 2022 Gemba Advantage
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-*/
-
-package skip
-
-import (
-	"github.com/gembaadvantage/uplift/internal/context"
-	"github.com/gembaadvantage/uplift/internal/middleware"
-)
-
-// Action defines a function that supports the skipping of future middleware
-// in the chain. Action is defined with the same method signature as task.Runner
-// Skip(), allow taks to be transparently wrapped
-type Action func(ctx *context.Context) bool
-
-// Running checks if a skip condition is satisfied before invoking the action itself.
-// When used within a middleware chain, any wrap actions will be skipped if the
-// skip condition resolves to true
-func Running(skipAct Action, act middleware.Action) middleware.Action {
-	return func(ctx *context.Context) error {
-		if skipAct(ctx) {
-			return nil
-		}
-		return act(ctx)
-	}
-}
