+++ internal/task/runner_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2022 Gemba Advantage
+Copyright (c) 2023 Gemba Advantage
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -20,11 +20,31 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-package middleware
+package task_test
 
-import "github.com/gembaadvantage/uplift/internal/context"
+// TODO: Execute -> check that task is executed
+// TODO: Execute -> check that task is skipped
 
-// Action defines a function that allows middleware to easily be chained.
-// Action is defined with the same method signature as task.Runner Run(),
-// allowing tasks to be transparently wrapped more easily
-type Action func(ctx *context.Context) error
+/*
+func TestRunning(t *testing.T) {
+	Running(func(ctx *context.Context) bool {
+		return true
+	}, func(ctx *context.Context) error {
+		assert.Fail(t, "action should be skipped")
+		return nil
+	})(&context.Context{})
+}
+
+func TestRunning_NoSkip(t *testing.T) {
+	exec := false
+
+	Running(func(ctx *context.Context) bool {
+		return false
+	}, func(ctx *context.Context) error {
+		exec = true
+		return nil
+	})(&context.Context{})
+
+	assert.True(t, exec)
+}
+*/
